更新2020/04/17

+++ libhttp/source/http-upload.c
@@ -28,12 +28,12 @@ static const char* http_header_attr_token(const char* source, const char* token,
 	return source + 1 + strspn(source + 1, WHITESPACE);
 }
 
-static int http_header_attr_value(const char* source, char* value, size_t bytes)
+static int http_header_attr_value(const char* source, char* value, int bytes)
 {
-	size_t n;
+	int n;
 	
-	source += strspn(source, "\'\"");
-	n = strcspn(source, "; \t\r\n\'\"");
+	source += (int)strspn(source, "\'\"");
+	n = (int)strcspn(source, "; \t\r\n\'\"");
 	if (bytes > n || 0 == bytes)
 	{
 		memcpy(value, source, n);
